// src/lib/types.ts
export interface Review {
  id: number;
  servicePublicId?: string;
  providerUid?: string;
  serviceTitle?: string;
  userName: string | null;
  userAvatar?: string | null;
  userUid?: string | null;
  imageUrl?: string | null;
  rating: number;
  text: string | null;
  date: string;
  autoGenerated?: boolean;
  ownerReply?: string | null;
  likesCount?: number;
}

export interface ProviderProfile {
  uid: string;
  name: string;
  avatar: string | null;
  responseRate: string;
  joinedDate?: string;
  description?: string;
  reviewsCount?: number;
  rating?: number;
  reviews?: Review[];
  isPremium?: boolean;
  isVerified?: boolean;
  deleted?: boolean;
  joinedAt?: string | null;
  avgRating?: number;
  zdjecieTla?: string | null;
  servicesCount?: number;
  facebook?: string | null;
  instagram?: string | null;
  tiktok?: string | null;
  website?: string | null;
  imie?: string | null;
  nazwisko?: string | null;
  profilowe?: string | null;
  joinedYear?: number | null;
  city?: string | null;
}

export interface Service {
  publicId: string;
  title: string;
  description: string;
  price: string;
  priceUnit: string;
  rating: number;
  distance: string;
  city: string;
  location?: { lat: number; lng: number };
  category: string;
  radius: number;
  type: 'offer' | 'request';
  isRemote: boolean;
  provider: ProviderProfile;
  image: string;
  images?: string[];
  isOnline?: boolean;
  deliveryTime?: string;
  durationMinutes?: number;
  bookings?: number;
  isFavorite?: boolean;
  views?: number;
  createdAt?: string;
  phoneNumber?: string;
  address?: string;
  addressLat?: number;
  addressLng?: number;
}